Define a basic node implementation
This is likely not fully correct, especially on the `handle_message`
logic, but I will tolerate this for now.

It may present me the opportunity to write some tests to demonstrate the
incorrectness right now (if I manage to read the chapter on testing from
"Real World OCaml" in time).