Add basic scaffolding (#1)
Establishes basic types, a simulator harness

===

* Design Messages and State type interfaces

The following interfaces have also been implemented:
1. types.ml
2. message.ml

This should offer a good foundation on which we can build our subsystems.

* Define a basic node implementation

This is likely not fully correct, especially on the `handle_message`
logic, but I will tolerate this for now.

It may present me the opportunity to write some tests to demonstrate the
incorrectness right now (if I manage to read the chapter on testing from
"Real World OCaml" in time).

* Lay groundworks for a sound simulation/event bus

This is incomplete at the moment, the note diffs should show the
improved mental/conceptual model on how the different subsystems should
be working together.

* Partial wire-up

* Fix type mismatches

Using generated code was a real time-sink, there were too many
compilation issues to hunt down. Wasted roughly about a day on this :/

* Minor change: add simulation entrypoint

* Fix typing issues @ simulator

* Minor cleanups