[FIX] Snapshot bef `drain`, avoid infinite echoes
The simulation was wonky because my echoed responses would just end up
infinitely echoing.

The problem was that the drain function would do a while loop and exit
only when the message queue would be empty. However, the responses from
nodes would fillup this queue and so the queue would never be empty and
hence -- infinite echoing.

The fix was to just snapshot the batch of messages so that a fixed
number of messages would be dispatched when we call `drain`.

So far we've been "manually" trying to simulate "ticks" -- the next step
is to add in time simulation!!