Ensure simulation using `event_bus` works (#2)
* Wire up msg broadcast,verify callback registration

Trying to make sure the event stream is easy to use.

* Keep node subs as a Hashtbl, not list

* Use a nested Hashtable; allow Node to respond

* [FIX] Snapshot bef `drain`, avoid infinite echoes

The simulation was wonky because my echoed responses would just end up
infinitely echoing.

The problem was that the drain function would do a while loop and exit
only when the message queue would be empty. However, the responses from
nodes would fillup this queue and so the queue would never be empty and
hence -- infinite echoing.

The fix was to just snapshot the batch of messages so that a fixed
number of messages would be dispatched when we call `drain`.

So far we've been "manually" trying to simulate "ticks" -- the next step
is to add in time simulation!!

-----

This is what a simulation run looks like at the moment, primitive but correct!

```
ॐ ❯ dune exec bin/simulation.exe
Node 1 subscribed to topics Coordination, Suggestion
Node 2 subscribed to topics Coordination, Suggestion

-----------------------------------------------------------------------
 Topic                  | Subscribers | Published | Delivered | Queued
-----------------------------------------------------------------------
 Suggestion             |           2 |         0 |         0 |      0
 Coordination           |           2 |         0 |         0 |      1
-----------------------------------------------------------------------
--- Draining the queue of 1 items from batch start
>>> Rcv @ echo node 2:
        (PermissionRequest(meta((id 00000000-0000-4000-8000-000000000000)(timestamp 1761874896.6684239)(topic Coordination)))(from 1)(proposal((seq 1)(node 1)))(value foo))
 Node 2 just enqueued a response
Node 2 changed state from Echo to Idle
>>> Rcv @ echo node 1:
        (PermissionRequest(meta((id 00000000-0000-4000-8000-000000000000)(timestamp 1761874896.6684239)(topic Coordination)))(from 1)(proposal((seq 1)(node 1)))(value foo))
 Node 1 just enqueued a response
Node 1 changed state from Echo to Idle

-----------------------------------------------------------------------
 Topic                  | Subscribers | Published | Delivered | Queued
-----------------------------------------------------------------------
 Suggestion             |           2 |         0 |         0 |      0
 Coordination           |           2 |         1 |         2 |      2
-----------------------------------------------------------------------
--- Draining the queue of 2 items from batch start
Bro I'm idle - Node2
Bro I'm idle - Node1
Bro I'm idle - Node2
Bro I'm idle - Node1

-----------------------------------------------------------------------
 Topic                  | Subscribers | Published | Delivered | Queued
-----------------------------------------------------------------------
 Suggestion             |           2 |         0 |         0 |      0
 Coordination           |           2 |         3 |         6 |      0
-----------------------------------------------------------------------

```