[temp] Put quorum within msgs
I wanna move onto time simulation next, this is an incomplete thing for
now and that's okay:
1. need to use simulation topic, adjust quorum state based on simulation
commands that the simulator sends -- this is part of the simulation
system controls piece

----

This is a temporary implementation because I treat the quorum as
simulation / config information.

I think it should be local to the node's state, because it's like other computers that kind of know who else is their peer in their network.

This will likely drift (and in fact, there's a need to have a consensus on what the number of peers are). The SOT for this can be the Simulator, we can communicate that over a particular message of its own (for quorum edits or config edits).

This can happen over a simulation topic for all the simulation events. For now, let's just manually do it, we can do this @ every request message, and put it part of the message payload.

The actual quorum can just be mutable field within the node's state (ref). In the future, we can use the simulation topic to adjust, so the master / simulator will be the SOT on what the quorum is.

- [ ] implmementing the =is_quorum_reached=

  this is kind of paxos dependent. I know that the responding stuff works, so I'll come back to this later when implementing the simple paxos algo.