Create CLI cmds; refactor to use `assertion` type
Assertion and Promise go hand in hand, it's some basic types that the
rest of the system uses. `assertion` acts as the basic data unit for
paxos processes. The choice of name clashes with the programming
language use of the term "assertion", but I hope it's clear why I think
it's still a good name: peer nodes in a distributed system *assert* what
they think the value should be, and the paxos protocol is to achieve
consensus on such an assertion.

Thank god for the compiler -- the refactors for this commit were basically just big-bangs.