Support node partitions; use node & bus registries
This is a such a large commit, it was messy to move away from the
singleton event bus and be able to define partitions with their own bus.

The value that we get from implementing this is that now we can simulate
network partitioning as well as healing of network partitions.

Other things done:

1. the input jsons can omit the optional fields for the config

2. support registering and deregistering routines for node w.r.t a bus
and the more upstream versions of those routines